Add and test game- & score- counting functionality
Changed return values of playRound() to return a value representing
who won (logging the win message instead of returning it). Used this
to tally total wins for each player.
Added conditional statement in game() to return final score using said
return values.
Added a counter to game() to keep track of which game it was and added
it to the prompt() message to tell user.